Changed time values on long due to the need to work with timestamp in the database

+++ src/main/java/com/example/postebintest/data/ExpirationTime.java
@@ -15,18 +15,18 @@ public enum ExpirationTime {
   ONE_MONTH,
   NO_LIMITS;
 
-  private static final Map<Integer, ExpirationTime> valuesMap = Map.of(
-    10, TEN_MINUTES,
-    60, ONE_HOUR,
-    180, THREE_HOURS,
-    1440, ONE_DAY,
-    10080, ONE_WEEK,
-    44640, ONE_MONTH,
-    0, NO_LIMITS
+  private static final Map<Long, ExpirationTime> valuesMap = Map.of(
+    10L, TEN_MINUTES,
+    60L, ONE_HOUR,
+    180L, THREE_HOURS,
+    1440L, ONE_DAY,
+    10080L, ONE_WEEK,
+    44640L, ONE_MONTH,
+    0L, NO_LIMITS
   );
 
   @JsonCreator
-  public static ExpirationTime forValue(Integer value) {
+  public static ExpirationTime forValue(Long value) {
     ExpirationTime expirationTime = valuesMap.get(value);
     if (expirationTime != null) {
       return expirationTime;
@@ -35,8 +35,8 @@ public enum ExpirationTime {
   }
 
   @JsonValue
-  public Integer toValue() {
-    for (Map.Entry<Integer, ExpirationTime> entry : valuesMap.entrySet()) {
+  public Long toValue() {
+    for (Map.Entry<Long, ExpirationTime> entry : valuesMap.entrySet()) {
       if (entry.getValue() == this) {
         return entry.getKey();
       }
